Use braced syntax even for tuple and units, in clone_enum
Demonstrate the named-vs-unnamed-agnostic approach.

Unit, Named and Unnamed are all now handled by the named fields code,
using the `{ }` syntax which Rust allows for every `Data`.

The `match` statement is redundant, but removing it will involve a
deindent so is very textually invasive.  We'll do that in the next
commit.