Use `0:` for named-fields debug output of tuples
Change the output when the user requests debug printing of a tuple
struct in named fields syntax.

Previously we output `_0: ` (for example), which isn't really right:
Rust permits `0: `, and the field is actually called `0`.

This is a behavioural change, but I don't think Debug output is
regarded as semver-relevant, so this shouldn't be a breaking change.